+++ src/instruments/src/DCDU/elements/Checkerboard.tsx
@@ -9,27 +9,17 @@ type CheckerboardProps = {
     fill: string
 }
 
-function createCells(offsetUsed: boolean, x: number, y: number, xIndex: number, yIndex: number, lineCount: number, columnCount: number, cellWidth: number, color: string) {
+function createCells(x: number, y: number, yIndex: number, lineCount: number, cellWidth: number, color: string, even: number[], odd: number[]) {
     if (yIndex >= lineCount) {
         return <></>;
-    } if (xIndex >= columnCount) {
-        if (!offsetUsed) {
-            return createCells(true, x, y + cellWidth, 0, yIndex + 1, lineCount, columnCount, cellWidth, color);
-        }
-        return createCells(false, x, y + cellWidth, 0, yIndex + 1, lineCount, columnCount, cellWidth, color);
     }
 
-    let xOffset = xIndex * cellWidth;
-    if (offsetUsed) {
-        xOffset += cellWidth;
-    }
+    const coordinates = yIndex % 2 ? odd : even;
 
     return (
         <>
-            <rect x={x + xOffset} y={y} width={cellWidth} height={cellWidth} style={{ fill: color }} />
-            (
-            {createCells(offsetUsed, x, y, xIndex + 2, yIndex, lineCount, columnCount, cellWidth, color)}
-            )
+            {coordinates.map((coordinate) => <rect x={x + coordinate} y={y + yIndex * cellWidth} width={cellWidth} height={cellWidth} style={{ fill: color }} />)}
+            {createCells(x, y, yIndex + 1, lineCount, cellWidth, color, even, odd)}
         </>
     );
 }
@@ -39,11 +29,17 @@ export const Checkerboard: React.FC<CheckerboardProps> = memo(({ x, y, width, he
     const lines = Math.round(height / cellSize);
     const columns = Math.round(width / cellSize);
 
+    // prepare lookup tables for the cells
+    const evenCoordinates: number[] = [];
+    const oddCoordinates: number[] = [];
+    for (let i = 0; i < columns; i += 2) {
+        evenCoordinates.push(i * cellSize + cellSize);
+        oddCoordinates.push(i * cellSize);
+    }
+
     return (
         <g>
-            (
-            {createCells(false, x, y, 0, 0, lines, columns, cellSize, fill)}
-            )
+            {createCells(x, y, 0, lines, cellSize, fill, evenCoordinates, oddCoordinates)}
         </g>
     );
 });